% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trace_sahul.R
\name{classify_TraCESahul}
\alias{classify_TraCESahul}
\title{Classify the input files based on whats provided}
\usage{
classify_TraCESahul(files)
}
\description{
This function classifies the input files into a number of different categories
depending on whats passed which is then used to switch the function used for
importing the data. This function is no exported, and should not ever need to
be called directly by the user.
}
