% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_CHELSA.R
\name{download_CHELSA}
\alias{download_CHELSA}
\title{Source daily CHELSA climate rasters}
\usage{
download_CHELSA(
  x,
  var,
  dir,
  template,
  algo = "cubicspline",
  mask = TRUE,
  convert = TRUE,
  ...
)
}
\arguments{
\item{x}{Date. Must be a \code{\link[base]{Date}} class object/vector}

\item{var}{Character. CHELSA variable name. Must be one of \code{pr}, \code{tasmax}, or \code{tasmin}.}

\item{dir}{Character. Base output directory. A subdirectory named \code{var} is
created if needed.}

\item{template}{\code{\link[terra:rast]{SpatRast}} grid for resampling and cropping. Should be a \emph{TraCE-Sahul} grid.}

\item{algo}{Character. Resampling method passed to \code{\link[terra]{resample}}.}

\item{mask}{Logical. If \code{TRUE} mask to the non-NA values of the template raster}

\item{convert}{Logical. If \code{TRUE}, apply variable-specific unit conversion.}

\item{...}{Additional arguments passed to \code{\link[terra]{writeRaster}}.}
}
\value{
\code{file.path} to the cropped and resampled raster
}
\description{
Downloads a monthly CHELSA v2.1 GeoTIFF for a given date and variable via GDAL's
virtual file system (\href{https://gdal.org/en/stable/user/virtual_file_systems.html#network-based-file-systems}{\emph{vsicurl}}),
reads only a buffered window around \code{template}, resamples to the template grid, optionally masks, converts units, and
writes the result to disk.
}
\details{
The crop window is hard coded to the extent of \code{template}, buffered by 50 km.
Output files are written to \code{file.path(dir, var)} and named using the source
layer name (for example, \code{"CHELSA_pr_01_1980_V.2.1.tif"}).

Note that the final raster is resampled to match the exact extent of the template raster,
but data is read/downloaded for the extent of the template + 50km.

If \code{convert = TRUE}:

\code{tasmin} and \code{tasmax} are converted from \emph{Kelvin} to \emph{Celsius}, \code{pr} is
converted from \emph{mm/month} to \emph{kg m-2 s-1}.
}
\examples{
\dontrun{
library(terra)

template <- rast(
  ext(105, 161.25, -50, 10),
  resolution = 0.05,
  crs = "EPSG:4326"
)

r <- download_CHELSA(
  x = as.Date("1980-01-01"),
  var = "pr",
  dir = tempdir(),
  template = template,
  algo = "cubicspline",
  convert = TRUE,
  mask = FALSE,
  overwrite = TRUE
)

r
}
}
