% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trace_data.R
\name{download_trace_cmip6_data}
\alias{download_trace_cmip6_data}
\title{Download and extract the example TraCE-Sahul CMIP6 dataset}
\usage{
download_trace_cmip6_data(destdir, ...)
}
\arguments{
\item{destdir}{Character. Path to the directory where the data should be downloaded and extracted.}

\item{...}{Optional parameters to \code{\link[utils]{download.file}}. \code{mode = "wb"} is fixed.}
}
\value{
Invisibly returns the path to the extracted directory.
}
\description{
Downloads and extracts a file from box (~150 MB) which contains a spatial
subset of the \emph{TraCE-Sahul} CMIP6 dataset for the Arafura–Timor region
(125–135°E, 8–13°S) from 1900 to 2100 for 4 SSP scenarios. Will extract to a
folder called \emph{CMIP6} in the specified directory.
}
\note{
The timeout for the download is set at 30 minutes.
}
\examples{
\dontrun{
download_trace_data("~/Downloads")
}
}
