% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_foss}
\alias{ex_foss}
\title{Example fossil dataset}
\format{
A \code{\link[terra:vect]{SpatVect}} with columns:
\describe{
  \item{date_estimate}{Age estimate of fossil record}
  \item{ci_lower, ci_upper}{Lower and upper CI of fossil age}
}

A \code{\link[terra:vect]{PackedSpatVector}}
}
\source{
data-raw/ex_foss.R
}
\usage{
data(ex_foss)
}
\description{
A \emph{PackedSpatVector} object with locations and ages of fossils
for a virtual species. Use \code{\link[terra]{wrap}} to convert back to a
\emph{SpatVector}.
}
\examples{
\dontrun{
library(terra)
ex_foss
ex_foss <- terra::unwrap(ex_foss)
ex_foss
}
}
\keyword{datasets}
