% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTime.R
\name{setTime}
\alias{setTime}
\title{Set time index for TraCE-Sahul SpatRaster}
\usage{
setTime(x)
}
\arguments{
\item{x}{A filepath pointing to the \emph{TraCE-Sahul} dataset(s)}
}
\value{
A \emph{SpatRaster} with updated time index showing the relevant year

A \emph{data.table} in the \code{.GlobalEnv} named \code{TraCESahul_time_steps} with the relevant details for all timesteps for \code{x}
}
\description{
This function sets the \code{\link[terra:time]{time index}} of a SpatRaster
representing the full (or a subset) from the \emph{TraCE-Sahul} decadal
dataset. Internally the function reads a CSV of timesteps from
\code{inst/extdata/TraCE-Sahul_timesteps.csv}.
}
\details{
The function will stop if \code{x} is not a file.path.

The function \strong{should} will with the subsets of the decadal data
e.g \emph{TraCE_22ka_downscaled_pr_decadal_21k_1500CE_biascorr_000001.nc}
provided that they haven't been renamed!
}
\note{
There are no checks done to ensure that the dataset passed is decadal data
from \emph{TraCE-Sahul}.
}
\examples{
\dontrun{
library(terra)
fn <- "TraCE_22ka_downscaled_pr_decadal_21k_1500CE_biascorr_000001.nc"
climData <- setTime(fn)
climData
}

}
