% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TraCESahul_timesteps}
\alias{TraCESahul_timesteps}
\title{Timesteps for TraCE-Sahul}
\format{
A \code{\link[data.table:data.table]{DataTable}} of 31,740 rows and
11 columns
}
\source{
data-raw/ex_foss.R
}
\usage{
data(TraCESahul_timesteps)
}
\description{
A data.table object containing the timesteps for each layer in the
TraCE-Sahul downscaled climate data.
}
\examples{
\dontrun{
library(data.table)
TraCESahul_timesteps
}
}
\keyword{datasets}
