% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_suit}
\alias{true_suit}
\title{Example habitat suitability}
\format{
A \code{\link[terra:rast]{PackedSpatRaster}}
}
\source{
data-raw/ex_foss.R
}
\usage{
data(true_suit)
}
\description{
A \emph{PackedSpatRaster} object with "true" contemporary habitat suitability
for a virtual species. Use \code{\link[terra]{wrap}} to convert back to a
\code{\link[terra:rast]{SpatRast}}.
}
\examples{
\dontrun{
library(terra)
true_suit
true_suit <- terra::unwrap(true_suit)
true_suit
}
}
\keyword{datasets}
