% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trace_data.R
\name{download_trace_data}
\alias{download_trace_data}
\title{Download and extract the example TraCE-Sahul dataset}
\usage{
download_trace_data(destdir, ...)
}
\arguments{
\item{destdir}{Character. Path to the directory where the data should be downloaded and extracted.}

\item{...}{Optional parameters to \code{\link[utils]{download.file}}. \code{mode = "wb"} is fixed.}
}
\value{
Invisibly returns the path to the extracted directory.
}
\description{
Downloads and extracts a file from box (~2 GB) which contains a spatial
subset of the \emph{TraCE-Sahul} dataset for the Arafura–Timor region
(125–135°E, 8–13°S) from 21k BP to 1990 CE. Will extract to a folder called
\emph{Trace-Sahul} in the specified directory.
}
\note{
The timeout for the download is set at 30 minutes.
}
\examples{
\dontrun{
download_trace_data("~/Downloads")
}
}
