% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trace_sahul.R
\name{import_TraCESahul}
\alias{import_TraCESahul}
\title{Imports the TraCE-Sahul dataset}
\usage{
import_TraCESahul(files)
}
\arguments{
\item{files}{A character vector of filepaths pointing to the \emph{TraCE-Sahul} dataset(s).}
}
\value{
A \emph{SpatRaster} with updated \code{\link[terra]{time}} index showing the relevant year, and \code{\link[terra]{depth}} representing the month

A \emph{data.table} in the \code{.GlobalEnv} named \code{TraCESahul_time_steps} with the relevant details for all timesteps for \code{files}
}
\description{
This function returns a \code{\link[terra:rast]{SpatRast}} representing the
full (or a subset) \emph{TraCE-Sahul} dataset. Internally the
function reads a CSV of timesteps from \code{inst/extdata/TraCE-Sahul_timesteps.csv},
and exports it to the \code{.GlobalEnv}. This data.table is \emph{only} relevant
to the decadal data.

If the monthly data from 1500-1989 is also provided, it will be appended.
}
\details{
The function will stop if \code{x} is not a file.path or series of file.paths.

The function will work with the subsets of the decadal data
e.g \emph{TraCE_22ka_downscaled_pr_decadal_21k_1500CE_biascorr_01.nc}
provided that they haven't been renamed!
}
\note{
The function will fail if the \emph{TraCE-Sahul} datasets have been renamed.

filepaths must be given in order (e.g. 01.nc, 02.nc, 03.nc, 04.nc, 05.nc, 06.nc, 1500_1990_biascorr.nc)
}
\examples{
\dontrun{
fn <- c("TraCE_22ka_downscaled_pr_decadal_21k_1500CE_biascorr_06.nc",
        "TraCE_22ka_downscaled_pr_1500_1990_biascorr.nc")
        sahul_pr <- import_TraCESahul(fn)
sahul_pr
fn <- c("TraCE_22ka_downscaled_tasmax_decadal_21k_1500CE_biascorr_06.nc",
        "TraCE_22ka_downscaled_tasmax_1500_1990_biascorr.nc")
sahul_tasmax <- import_TraCESahul(fn)
sahul_tasmax
fn <- c("TraCE_22ka_downscaled_tasmin_decadal_21k_1500CE_biascorr_06.nc",
        "TraCE_22ka_downscaled_tasmin_1500_1990_biascorr.nc")
sahul_tasmin <- import_TraCESahul(fn)
sahul_tasmin
}
}
