% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_trace_sahul.R
\name{import_TraCESahul}
\alias{import_TraCESahul}
\title{Imports the TraCE-Sahul dataset}
\usage{
import_TraCESahul(files, aoi = NULL)
}
\arguments{
\item{files}{A character vector of filepaths pointing to the
\emph{TraCE-Sahul} dataset(s).}

\item{aoi}{an \code{\link[terra]{ext}} object or four parameter vector
(\code{c(xmin,xmax,ymin,ymax)}) passed to \code{\link[terra]{rast}} to
spatially subset the \emph{TraCE-Sahul} data. Default = \code{NULL} which will
return the entire spatial domain.}
}
\value{
A \emph{SpatRaster} with updated \code{\link[terra]{time}} index
showing the relevant year, and \code{\link[terra]{depth}} representing the month

A \emph{data.table} in the \code{.GlobalEnv} named \code{TraCESahul_time_steps} with the relevant details for all timesteps for \code{files}
}
\description{
This function returns a \code{\link[terra:rast]{SpatRast}} representing the
full (or a subset) \emph{TraCE-Sahul} dataset. Internally the
function reads a CSV of timesteps from \code{inst/extdata/TraCE-Sahul_timesteps.csv},
and exports it to the \code{.GlobalEnv}. This data.table is \emph{only} relevant
to the decadal data.

If the monthly data from 1500-1989 is also provided, it will be appended.
}
\details{
The function will stop if \code{x} is not a file.path or series of file.paths.

The function will work with the subsets of the decadal data
e.g \emph{TraCE_22ka_downscaled_pr_decadal_21k_1500CE_biascorr_01.nc}
provided that they haven't been renamed!
}
\note{
The function will fail if the \emph{TraCE-Sahul} datasets have been renamed.

filepaths must be given in order (e.g. 01.nc, 02.nc, 03.nc, 04.nc, 05.nc, 06.nc, 1500_1990_biascorr.nc)
}
\examples{
\dontrun{
vars <- c("pr", "tasmax", "tasmin")
load_tracesahul <- function(base, var) {
 f1 <- file.path(base, var,
                 sprintf("TraCE_22ka_downscaled_\%s_decadal_21k_1500CE_biascorr.nc",
                 var))
 f2 <- file.path(base, var,
                 sprintf("TraCE_22ka_downscaled_\%s_1500_1990_biascorr.nc",
                 var))
import_TraCESahul(c(f1, f2))
}
sahul <- lapply(vars, function(v) load_tracesahul(base_dir, v))
names(sahul) <- vars
sahul
}
}
