% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_TraCESahul.R
\name{summarise_TraCESahul}
\alias{summarise_TraCESahul}
\title{summarise_TraCESahul}
\usage{
summarise_TraCESahul(x, type = "annual", sumfun = "mean", ...)
}
\arguments{
\item{x}{a \code{\link[terra:rast]{SpatRast}}. \strong{Must} be created using \code{\link{import_TraCESahul}}.}

\item{type}{which timesteps to summarise over. Can be seasonal or annual. Default: "annual".}

\item{sumfun}{summary function to be applied. See \code{\link[terra]{tapp}} for details. Default: "mean".}

\item{...}{additional argument to \code{\link[terra]{tapp}}}
}
\value{
a \code{\link[terra:rast]{SpatRast}} summarised over the chosen value
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\dontrun{
pr_summary <- summarise_TraCEShaul(
x = pr_chunks[[1:120]], # first 120 layers only
type = "seasonal",
sumfun = "mean")
}
}
